+++ web/usuaris/src/components/organisms/FiltersWidget.tsx
@@ -4,7 +4,7 @@ import { FlexBox, Search, dimensions } from '@itacademy/ui'
 import { useTranslation } from 'react-i18next'
 import { DateRange, StatusDropdown } from '../molecules'
 import { ItineraryDropdown } from '../molecules/ItineraryDropdown'
-import { type TFilters, TItinerary, type TStatus } from '../../types'
+import { type TFilters, TItinerary, UserStatus } from '../../types'
 
 const FiltersContainer = styled(FlexBox)`
   width: 100%;
@@ -12,53 +12,39 @@ const FiltersContainer = styled(FlexBox)`
 
 type TFiltersWidget = {
   filters: TFilters
-  handleFilters: (filters: TFilters) => void
+  setFilters: (filters: TFilters) => void
 }
 
-export const FiltersWidget: FC<TFiltersWidget> = ({
-  filters,
-  handleFilters,
-}) => {
+export const FiltersWidget: FC<TFiltersWidget> = ({ filters, setFilters }) => {
   const { t } = useTranslation()
 
   const handleItinerary = (itinerary: TItinerary | undefined) => {
-    if (itinerary !== undefined) {
-      handleFilters({ ...filters, itinerarySlug: itinerary.slug })
-    } else {
-      const newFilters = { ...filters }
-      delete newFilters.itinerarySlug
-      handleFilters(newFilters)
-    }
+    setFilters({ ...filters, itinerarySlug: itinerary?.slug })
   }
 
-  const handleStatus = (selectedStatus: TStatus | undefined) => {
-    if (selectedStatus !== undefined) {
-      handleFilters({ ...filters, status: selectedStatus.id })
-    } else {
-      const newFilters = { ...filters }
-      delete newFilters.status
-      handleFilters(newFilters)
-    }
+  const handleStatus = (selectedStatus: UserStatus | undefined) => {
+    setFilters({ ...filters, status: selectedStatus })
   }
 
   const handleSearch = useCallback(
     (searchValue: string) => {
-      if (searchValue !== '') {
-        handleFilters({
-          ...filters,
-          name: searchValue,
-          dni: searchValue,
-        })
-      } else {
-        const newFilters = { ...filters }
-        delete newFilters.name
-        delete newFilters.dni
-        handleFilters(newFilters)
-      }
+      setFilters({
+        ...filters,
+        name: searchValue,
+        dni: searchValue,
+      })
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    []
+    [filters]
   )
+
+  const handleDates = (startDate: Date | null, endDate: Date | null) => {
+    setFilters({
+      ...filters,
+      startDate: startDate?.toISOString(),
+      endDate: endDate?.toISOString(),
+    })
+  }
 
   return (
     <FiltersContainer
@@ -75,6 +61,7 @@ export const FiltersWidget: FC<TFiltersWidget> = ({
         placeholderEndDate={t('Hasta...')}
         dateFormat="dd/MM/yyyy"
         calendarLanguage={t('calendarLanguage')}
+        handleDates={handleDates}
       />
       <Search
         id="usersSearch"