+++ web/usuaris/src/types/index.tsx
@@ -1,5 +1,5 @@
+export { UserStatus } from './types'
 export { type TFilters } from './types'
 export { type TItinerary } from './types'
-export { type TStatus } from './types'
 export { type TUpdatedUser } from './types'
 export { type TUserData } from './types'